requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==3.6.4.0
blinker==1.7.0
Brotli==1.1.0
bs4==0.0.2
celery==5.2.7
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
Django==5.1
exceptiongroup==1.2.2
fake-useragent==1.5.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.8
kaitaistruct==0.10
kombu==5.4.0
outcome==1.3.0.post0
packaging==24.1
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydivert==2.1.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
PySocks==1.7.1
python-dotenv==0.21.1
pytz==2024.1
redis==5.0.8
requests==2.32.3
selenium==4.24.0
selenium-wire==5.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
zstandard==0.23.0
